make general updates - split up tutorial so deploying a contract isn't dependent on sending a transaction - update so errors and warnings don't show in terminal as user tries out each function/section
