+++ .snippets/code/ethers-rust/deploy-contract/main.rs
@@ -1,3 +1,4 @@
+use ethers::providers::{Provider, Http};
 use ethers::{utils, prelude::*};
 use ethers_solc::Solc;
 use std::{path::Path, sync::Arc};
@@ -6,49 +7,17 @@ type Client = SignerMiddleware<Provider<Http>, Wallet<k256::ecdsa::SigningKey>>;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let provider: Provider<Http> = Provider::<Http>::try_from("http://localhost:9933")?;
+    let provider: Provider<Http> = Provider::<Http>::try_from("https://rpc.api.moonbase.moonbeam.network")?; // Change to correct network
     // Do not include the private key in plain text in any produciton code. This is just for demonstration purposes
-    let wallet: LocalWallet = "99b3c12287537e38c90a9219d4cb074a89a16e9cdb20bf85728ebd97c343e342"
+    let wallet: LocalWallet = "PRIVATE KEY OF YOUR FROM ADDRESS"
         .parse::<LocalWallet>()?
         .with_chain_id(Chain::MoonbeamDev);
     let client = SignerMiddleware::new(provider.clone(), wallet.clone());
 
-    send_transaction(&client).await?;
-    print_balances(&provider).await?;
     let addr = compile_deploy_contract(&client).await?;
     read_number(&client, &addr).await?;
     increment_number(&client, &addr).await?;
     read_number(&client, &addr).await?;
-
-    Ok(())
-}
-
-// Print the balance of a wallet
-async fn print_balances(provider: &Provider<Http>) -> Result<(), Box<dyn std::error::Error>> {
-    let address = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045".parse::<Address>()?;
-    let balance = provider.get_balance(address, None).await?;
-
-    println!("{} has {}", address, balance);
-    Ok(())
-}
-
-
-// Sends some native currency
-async fn send_transaction(client: &Client) -> Result<(), Box<dyn std::error::Error>> {
-    let address_from = "0x6Be02d1d3665660d22FF9624b7BE0551ee1Ac91b".parse::<Address>()?;
-    let address_to = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045".parse::<Address>()?;
-
-    println!(
-        "Beginning transfer of 1 native currency from {} to {}.",
-        address_from, address_to
-    );
-    let tx = TransactionRequest::new()
-        .to(address_to)
-        .value(U256::from(utils::parse_ether(1)?))
-        .from(address_from);
-    let tx = client.send_transaction(tx, None).await?.await?;
-
-    println!("Transaction Receipt: {}", serde_json::to_string(&tx)?);
 
     Ok(())
 }
